pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "insight"
version = "0.1.0"
description = "Better Comments"
authors = [{ name = "Russell Power", email = "russell.power@gmail.com" }]
dependencies = [
    "PyYAML",
    "dominate",
    "fastapi",
    "psycopg2-binary",
    "itsdangerous",
    "openai",
    "python-dotenv",
    "python-multipart",
    "sqlalchemy",
    "requests",
    "uvicorn",
    "pytest>=8.3.2",
    "bs4>=0.0.2",
    "apscheduler>=3.10.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["src"]
